requirements.txt
django >4.2,<4.3
Pillow
python-decouple
djangorestframework
markdown
django-filter